third_party/blink/web_tests/virtual/interest-for-no-partial-interest/external/wpt/html/semantics/the-button-element/interest-for/interestfor-partial-interest.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] Partial interest timing should not be observable
  assert_equals: Interest should be partial (before hot key) expected true but got false
Harness: the test ran to completion.
